uploadGallery, main_routes, createProduct, createVariant,transaction

+++ module/user/transport/gin/delete_user.go
@@ -4,6 +4,7 @@ import (
 	"OpenMarket/common"
 	"OpenMarket/component/appctx"
 	userbusiness "OpenMarket/module/user/business"
+	srrepository "OpenMarket/module/user/repository"
 	userstorage "OpenMarket/module/user/storage"
 	"net/http"
 	"strconv"
@@ -19,8 +20,19 @@ func DeleteUser(appCtx appctx.AppContext) gin.HandlerFunc {
 		if err != nil {
 			panic(err)
 		}
+
+		u, ok := c.MustGet(common.CurrentUser).(common.Requester)
+		if !ok {
+			panic(common.ErrUnauthorized(nil))
+		}
+
+		if u.GetUserId() != id {
+			panic(common.ErrPermission("you don't have permission to delete this user", nil))
+		}
 		store := userstorage.NewSQLStore(db)
-		biz := userbusiness.NewDeleteUserBusiness(store)
+		frepo := srrepository.NewFindUserRepo(store)
+		drepo := srrepository.NewDeleteUserRepo(store)
+		biz := userbusiness.NewDeleteUserBusiness(frepo, drepo)
 		if err := biz.DeleteUser(c.Request.Context(), id); err != nil {
 			panic(err)
 		}
